requirements.txt
streamlit
pandas
reportlab
xlsxwriter
Pillow
google-api-python-client
google-auth

